#1247 Move commentable_mixin.py up one directory as it's used in both nodes and symbols.
